requirements.txt
south
django-registration
django-crispy-forms